# docs/core/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageListResponse": {
        "description": "Paginated image list.",
        "properties": {
          "images": {
            "description": "List of image metadata",
            "items": {
              "$ref": "#/components/schemas/ImageMetadata"
            },
            "title": "Images",
            "type": "array"
          },
          "limit": {
            "description": "Pagination limit",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Pagination offset",
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total number of images matching filters",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "images",
          "total",
          "limit",
          "offset"
        ],
        "title": "ImageListResponse",
        "type": "object"
      },
      "ImageMetadata": {
        "description": "Metadata for a stored SSTV image.",
        "properties": {
          "callsign": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Operator callsign",
            "title": "Callsign"
          },
          "direction": {
            "description": "Receive or transmit",
            "pattern": "^(rx|tx)$",
            "title": "Direction",
            "type": "string"
          },
          "filepath": {
            "description": "Absolute path to image file",
            "title": "Filepath",
            "type": "string"
          },
          "frequency_hz": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dial frequency the image was received on (RX only)",
            "title": "Frequency Hz"
          },
          "height": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Image height in pixels (null if file unreadable)",
            "title": "Height"
          },
          "id": {
            "description": "Unique image identifier",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SSTVMode"
              },
              {
                "type": "null"
              }
            ],
            "description": "SSTV mode used (null when unknown)"
          },
          "snr_db": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signal quality (RX only)",
            "title": "Snr Db"
          },
          "thumbnail_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path to thumbnail image",
            "title": "Thumbnail Path"
          },
          "timestamp": {
            "description": "Capture/transmit timestamp (UTC)",
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Image width in pixels (null if file unreadable)",
            "title": "Width"
          }
        },
        "required": [
          "id",
          "filepath",
          "direction",
          "timestamp"
        ],
        "title": "ImageMetadata",
        "type": "object"
      },
...
      "SSTVMode": {
        "description": "Supported SSTV modes.",
        "enum": [
          "MartinM1",
          "MartinM2",
          "ScottieS1",
          "ScottieS2",
          "ScottieDX",
          "Robot36",
          "Robot72",
          "PD90",
          "PD120",
          "PD180",
          "PD240",
          "WraaseSC2_180"
        ],
        "title": "SSTVMode",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "SSTV (Slow-Scan Television) encode/decode engine",
    "title": "SSTeVe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/images": {
      "get": {
        "description": "List images with pagination and optional filtering.\n\nSupports filtering by:\n- Direction: \"rx\" (received) or \"tx\" (transmitted)\n- Mode: SSTV mode (e.g., \"MartinM1\", \"ScottieS1\")\n- Callsign: Operator callsign\n\nImages are returned in reverse chronological order (newest first).\n\nQuery Parameters:\n    limit: Number of images per page (1-100, default: 20)\n    offset: Pagination offset (default: 0)\n    direction: Filter by \"rx\" or \"tx\"\n    mode: Filter by SSTV mode\n    callsign: Filter by operator callsign",
        "operationId": "list_images_api_v1_images_get",
        "parameters": [
          {
            "description": "Number of images per page",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Number of images per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Pagination offset",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Pagination offset",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Filter by direction",
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(rx|tx)$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by direction",
              "title": "Direction"
            }
          },
          {
            "description": "Filter by SSTV mode",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SSTVMode"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by SSTV mode",
              "title": "Mode"
            }
          },
          {
            "description": "Filter by callsign",
            "in": "query",
            "name": "callsign",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by callsign",
              "title": "Callsign"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Images",
        "tags": [
          "Images",
          "images"
        ]
      }
    },
    "/api/v1/images/{image_id}": {
      "get": {
        "description": "Get metadata for a specific image.\n\nReturns detailed information about the image including:\n- File path and dimensions\n- SSTV mode used\n- Direction (RX/TX)\n- Signal quality metrics (for RX images)\n- Timestamp and callsign\n\nRaises:\n    404 Not Found: If image doesn't exist",
        "operationId": "get_image_api_v1_images__image_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "image_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Image Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Image",
        "tags": [
          "Images",
          "images"
        ]
      }
    },
...
  }
}